api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
boto3==1.34.19
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.2.0
python-dotenv==1.0.0
httpx==0.26.0
psutil==5.9.8